by module, split happy path spec file into 3
