+++ tests/api/springboot/account.spec.ts
@@ -1,0 +1,41 @@
+// tests/api/springboot/account.spec.ts
+import { expect } from '@playwright/test';
+import { test } from '../../../fixtures/springboot-chained.fixture';
+import { expectError, expectOk } from '../../../services/apis/base-api-client';
+
+test.describe('Account 帳號管理', () => {
+	test('應該能建立新帳號', async ({
+		springbootApi,
+		accountFixtureDeletedAfterward,
+		newAccountData,
+	}) => {
+		const response = await springbootApi.createAccount(newAccountData);
+		const accountId = expectOk(response);
+
+		expect(typeof accountId).toBe('number');
+		accountFixtureDeletedAfterward.ids.push(accountId);
+	});
+
+	test('應該能取得帳號詳情', async ({ springbootApi, existingAccount }) => {
+		const response = await springbootApi.getAccount(existingAccount.id);
+		const data = expectOk(response);
+
+		expect.soft(data).toMatchObject({
+			name: existingAccount.name,
+			status: existingAccount.status,
+		});
+	});
+
+	test('應該能更新帳號', async ({ springbootApi, existingAccount, updateAccountData }) => {
+		const response = await springbootApi.updateAccount(updateAccountData(existingAccount));
+		expectOk(response);
+	});
+
+	test('應該能刪除帳號', async ({ springbootApi, existingAccount }) => {
+		const response = await springbootApi.deleteAccount(existingAccount.id);
+		expectOk(response);
+
+		const getResponse = await springbootApi.getAccount(existingAccount.id);
+		expectError(getResponse, 404);
+	});
+});
